moved update db tests to core tests
